+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/catchup/consumer/manager/DefaultTransactionalEventProcessor.java
@@ -1,12 +1,11 @@
 package uk.gov.justice.services.event.sourcing.subscription.catchup.consumer.manager;
 
+import javax.inject.Inject;
+import javax.transaction.Transactional;
 import uk.gov.justice.services.event.sourcing.subscription.manager.CatchupEventBufferProcessor;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.EventConverter;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.PublishedEvent;
 import uk.gov.justice.services.messaging.JsonEnvelope;
-
-import javax.inject.Inject;
-import javax.transaction.Transactional;
 
 import static javax.transaction.Transactional.TxType.REQUIRES_NEW;