+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/catchup/consumer/task/EventQueueConsumer.java
@@ -1,14 +1,13 @@
 package uk.gov.justice.services.event.sourcing.subscription.catchup.consumer.task;
 
+import java.util.Queue;
+import java.util.UUID;
+import javax.inject.Inject;
 import uk.gov.justice.services.event.sourcing.subscription.catchup.consumer.manager.CatchupEventProcessor;
 import uk.gov.justice.services.event.sourcing.subscription.catchup.consumer.manager.EventStreamConsumptionResolver;
 import uk.gov.justice.services.event.sourcing.subscription.catchup.consumer.manager.FinishedProcessingMessage;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.PublishedEvent;
 import uk.gov.justice.services.eventstore.management.commands.CatchupCommand;
-
-import javax.inject.Inject;
-import java.util.Queue;
-import java.util.UUID;
 
 public class EventQueueConsumer {