+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/error/StreamErrorRepository.java
@@ -1,6 +1,7 @@
 package uk.gov.justice.services.event.sourcing.subscription.error;
 
 import static javax.transaction.Transactional.TxType.MANDATORY;
+import static javax.transaction.Transactional.TxType.REQUIRED;
 
 import uk.gov.justice.services.event.buffer.core.repository.streamerror.StreamError;
 import uk.gov.justice.services.event.buffer.core.repository.streamerror.StreamErrorDetails;
@@ -11,11 +12,14 @@ import uk.gov.justice.services.jdbc.persistence.ViewStoreJdbcDataSourceProvider;
 
 import java.sql.Connection;
 import java.sql.SQLException;
+import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 import javax.inject.Inject;
 import javax.transaction.Transactional;
 
+@SuppressWarnings("java:S1192")
 public class StreamErrorRepository {
 
     @Inject
@@ -58,6 +62,26 @@ public class StreamErrorRepository {
         try (final Connection connection = viewStoreJdbcDataSourceProvider.getDataSource().getConnection()) {
             streamStatusErrorPersistence.unmarkStreamStatusAsErrored(streamId, source, componentName, connection);
             streamErrorPersistence.removeErrorForStream(streamId, source, componentName, connection);
+        } catch (final SQLException e) {
+            throw new StreamErrorHandlingException("Failed to get connection to view-store", e);
+        }
+    }
+
+    @Transactional(REQUIRED)
+    public List<StreamError> findAllByStreamId(final UUID streamId) {
+
+        try (final Connection connection = viewStoreJdbcDataSourceProvider.getDataSource().getConnection()) {
+            return streamErrorPersistence.findAllByStreamId(streamId, connection);
+        } catch (final SQLException e) {
+            throw new StreamErrorHandlingException("Failed to get connection to view-store", e);
+        }
+    }
+
+    @Transactional(REQUIRED)
+    public Optional<StreamError> findByErrorId(final UUID errorId) {
+
+        try (final Connection connection = viewStoreJdbcDataSourceProvider.getDataSource().getConnection()) {
+            return streamErrorPersistence.findByErrorId(errorId, connection);
         } catch (final SQLException e) {
             throw new StreamErrorHandlingException("Failed to get connection to view-store", e);
         }