+++ framework-rest-resources/framework-stream-rest-resources/src/test/java/uk/gov/justice/services/resources/FrameworkApplicationTest.java
@@ -2,11 +2,12 @@ package uk.gov.justice.services.resources;
 
 import java.util.Set;
 import org.junit.jupiter.api.Test;
+import uk.gov.justice.services.resources.rest.streams.StreamErrorsResource;
 import uk.gov.justice.services.resources.rest.streams.StreamsResource;
 
+import static org.hamcrest.CoreMatchers.hasItems;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.collection.IsIterableContainingInOrder.contains;
 
 class FrameworkApplicationTest {
 
@@ -16,7 +17,7 @@ class FrameworkApplicationTest {
     void validateConfiguredResourceClasses() {
         final Set<Class<?>> resourceClasses = frameworkApplication.getClasses();
 
-        assertThat(resourceClasses.size(), is(1));
-        assertThat(resourceClasses, contains(StreamsResource.class));
+        assertThat(resourceClasses.size(), is(2));
+        assertThat(resourceClasses, hasItems(StreamsResource.class, StreamErrorsResource.class));
     }
 }